Started Important Function
April 5, 2024

I've decided to start keeping track of my daily progress with these.
Majority of today's free time spent on learning how search algorithms work, nevertheless started on a function that allows to check for if the script's user is able to move in a direction with steps or slopes or just a flat surface, later I will make something to check if you can jump over the same thing. These functions will later be used to start trying to convert 3D space to "2D space" which can just use the most basic forms of pathfinding algorithms, not having to use 3D voxels.