Update BasicNPCAI.cs
Added code I made for an inefficient pathfinding algorithm that would use way too many raycasts in one frame. Can't really tell everything that was going through my brain here, but I think it could've worked. It was based off of trial and error and just stopping a path when a loop in it is detected and making it so the code wouldn't go there anymore, and it would keep trying until it did it right. This is wildly inefficient as the goal is to get the path right as few times as possible, and doing this method would force it to go through countless incorrect possibilities. From this point on, I'm making a new function that uses OverlapSphere and takes the corners of all the objects to see if the character can pass them to get to the goal.